pyproject.toml
[tool.poetry]
name = "opencommunity"
version = "0.1.0"
description = "Open Source API and data service that implements the HSDS"
authors = ["mathewtrivett <matt.trivett@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.5"
django-rest-marshmallow = "^4.0"
djangorestframework = "^3.9"
django = "^2.2"

[tool.poetry.dev-dependencies]
pytest-django = "^3.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
